requirements.txt
beautifulsoup4>=4.10.0
ffmpeg-python>=0.2.0
pathvalidate>=2.4.1
PySimpleGUI>=4.47.0
regex>=2021.8.28
requests>=2.26.0
